+++ src/pages/index.tsx
@@ -1,4 +1,5 @@
 import Layout from "../components/Layout";
+import NotFound from "./404NotFound";
 import Homepage from "./Homepage";
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 
@@ -6,9 +7,10 @@ export default function Pages() {
     return (
         <Router>
             <Layout>
-                <Routes>            
+                <Routes>      
+                    <Route path="*" element={<NotFound />} />      
                     <Route path="/" element={<Homepage />} />
-                    <Route path="/homepage" element={<Homepage />} />
+                    {/* <Route path="/homepage" element={<Homepage />} /> */}
                 </Routes>
             </Layout>
         </Router>